+++ src/main/java/br/com/nutriCenter/config/ModelMapperConfig.java
@@ -1,0 +1,21 @@
+/**
+ * 
+ */
+package br.com.nutriCenter.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author José Caio
+ *
+ */
+@Configuration
+public class ModelMapperConfig {
+	@Bean
+	public ModelMapper modelMapper() {
+		return new ModelMapper();
+	}
+
+}
