+++ src/main/java/br/com/nutriCenter/model/Paciente.java
@@ -1,19 +1,26 @@
 package br.com.nutriCenter.model;
+
 import java.util.Date;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
+
+import org.hibernate.validator.constraints.br.CPF;
+
 /**
  * @author José Caio
  *
  */
 
 @Entity
-@Table(name ="tb_paciente")
+@Table(name = "tb_paciente")
 public class Paciente extends Usuario {
 
 	private static final long serialVersionUID = 1L;
-	
+	@CPF
+	@Column(name = "CPF", nullable = false, unique = true)
+	private String cpf;
 	private String cidade;
 	private String rua;
 	private String bairro;
@@ -25,50 +32,100 @@ public class Paciente extends Usuario {
 		super.setNivelDeAcesso(1);
 	}
 
+	/**
+	 * @return the cpf
+	 */
+	public String getCpf() {
+		return cpf;
+	}
+
+	/**
+	 * @param cpf the cpf to set
+	 */
+	public void setCpf(String cpf) {
+		this.cpf = cpf;
+	}
+
+	/**
+	 * @return the cidade
+	 */
 	public String getCidade() {
 		return cidade;
 	}
 
+	/**
+	 * @param cidade the cidade to set
+	 */
 	public void setCidade(String cidade) {
 		this.cidade = cidade;
 	}
 
+	/**
+	 * @return the rua
+	 */
 	public String getRua() {
 		return rua;
 	}
 
+	/**
+	 * @param rua the rua to set
+	 */
 	public void setRua(String rua) {
 		this.rua = rua;
 	}
 
+	/**
+	 * @return the bairro
+	 */
 	public String getBairro() {
 		return bairro;
 	}
 
+	/**
+	 * @param bairro the bairro to set
+	 */
 	public void setBairro(String bairro) {
 		this.bairro = bairro;
 	}
 
+	/**
+	 * @return the estado
+	 */
 	public String getEstado() {
 		return estado;
 	}
 
+	/**
+	 * @param estado the estado to set
+	 */
 	public void setEstado(String estado) {
 		this.estado = estado;
 	}
 
+	/**
+	 * @return the dataCadastro
+	 */
 	public Date getDataCadastro() {
 		return dataCadastro;
 	}
 
+	/**
+	 * @param dataCadastro the dataCadastro to set
+	 */
 	public void setDataCadastro(Date dataCadastro) {
 		this.dataCadastro = dataCadastro;
 	}
 
+	/**
+	 * @return the dataUltimaConsulta
+	 */
 	public Date getDataUltimaConsulta() {
 		return dataUltimaConsulta;
 	}
 
+	/**
+	 * @param dataUltimaConsulta the dataUltimaConsulta to set
+	 */
 	public void setDataUltimaConsulta(Date dataUltimaConsulta) {
 		this.dataUltimaConsulta = dataUltimaConsulta;
 	}
