+++ src/test/java/br/com/nutriCenter/resourceTest/AdministradorResourceTeste.java
@@ -1,12 +1,10 @@
 package br.com.nutriCenter.resourceTest;
 
-import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 //import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-
 //import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -15,32 +13,21 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
-import org.mockito.BDDMockito.Then;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mysql.cj.x.protobuf.MysqlxCrud.Delete;
-import com.mysql.cj.x.protobuf.MysqlxCrud.Update;
 
 import br.com.nutriCenter.model.Administrador;
-import br.com.nutriCenter.model.Nutricionista;
 import br.com.nutriCenter.resource.AdministradorResource;
-import br.com.nutriCenter.resource.NutricionistaResource;
 import br.com.nutriCenter.services.AdministradorService;
-import br.com.nutriCenter.services.NutricionistaService;
-import groovy.transform.Undefined.EXCEPTION;
-import groovyjarjarpicocli.CommandLine.ExecutionException;
-import io.restassured.http.ContentType;
-import io.restassured.module.mockmvc.RestAssuredMockMvc;
 
 /**
  * @author Mateus Fernandes
@@ -77,7 +64,6 @@ public class AdministradorResourceTeste {
 		administrador.setSobreNome("Fernandes");
 		administrador.setEmail("mateus@email.com");
 		administrador.setGenero("Masculino");
-		administrador.setCpf("222-222-222-22");
 
 	}
 
@@ -139,7 +125,6 @@ public class AdministradorResourceTeste {
 		admin.setSobreNome(sobrenome);
 		admin.setEmail(email);
 		admin.setGenero(genero);
-		admin.setCpf(cpf);
 		
 		
 		return null;
